+++ src/helpers/string_helpers.ts
@@ -7,6 +7,7 @@ import { mastodon } from 'masto';
 import { MediaCategory } from '../types';
 
 export const DEFAULT_FONT_SIZE = 15;
+export const NULL = "<<NULL>>";
 export const GIFV = "gifv";
 export const IMAGE_EXTENSIONS = ["gif", "jpg", "jpeg", "png", "webp"];
 export const VIDEO_EXTENSIONS = ["mp4"];